+++ wemake_python_styleguide/visitors/ast/builtins.py
@@ -3,28 +3,16 @@
 import ast
 from collections import Counter, Hashable, defaultdict
 from contextlib import suppress
-from typing import (
-    ClassVar,
-    DefaultDict,
-    Iterable,
-    List,
-    Mapping,
-    Optional,
-    Sequence,
-    Tuple,
-    Type,
-    Union,
-)
+from typing import ClassVar, DefaultDict, Iterable, List, Sequence, Union
 
 import astor
 from typing_extensions import final
 
 from wemake_python_styleguide import constants
 from wemake_python_styleguide.compat.aliases import FunctionNodes
-from wemake_python_styleguide.logic import safe_eval, walk
+from wemake_python_styleguide.logic import safe_eval
 from wemake_python_styleguide.logic.naming.name_nodes import extract_name
 from wemake_python_styleguide.logic.operators import (
-    count_unary_operator,
     get_parent_ignoring_unary,
     unwrap_starred_node,
     unwrap_unary_node,
@@ -39,9 +27,6 @@ from wemake_python_styleguide.violations.best_practices import (
     WrongUnpackingViolation,
 )
 from wemake_python_styleguide.visitors import base, decorators
-
-_MeaninglessOperators = Mapping[int, Tuple[Type[ast.operator], ...]]
-_OperatorLimits = Mapping[Type[ast.unaryop], int]
 
 
 @final
@@ -105,115 +90,6 @@ class MagicNumberVisitor(base.BaseNodeVisitor):
 
 
 @final
-class UselessOperatorsVisitor(base.BaseNodeVisitor):
-    """Checks operators used in the code."""
-
-    _limits: ClassVar[_OperatorLimits] = {
-        ast.UAdd: 0,
-        ast.Invert: 1,
-        ast.Not: 1,
-        ast.USub: 1,
-    }
-
-    _meaningless_operations: ClassVar[_MeaninglessOperators] = {
-        # ast.Div is not in the list,
-        # since we have a special violation for it.
-        0: (ast.Mult, ast.Add, ast.Sub, ast.Pow),
-        # `1` and `-1` are different, `-1` is allowed.
-        1: (ast.Div, ast.Mult, ast.Pow),
-    }
-
-    def visit_Num(self, node: ast.Num) -> None:
-        """
-        Checks numbers unnecessary operators inside the code.
-
-        Raises:
-            UselessOperatorsViolation
-
-        """
-        self._check_operator_count(node)
-        self.generic_visit(node)
-
-    def visit_BinOp(self, node: ast.BinOp) -> None:
-        """
-        Visits binary operators.
-
-        Raises:
-            ZeroDivisionViolation
-
-        """
-        self._check_zero_division(node.op, node.right)
-        self._check_useless_math_operator(node.op, node.left, node.right)
-        self.generic_visit(node)
-
-    def visit_AugAssign(self, node: ast.AugAssign) -> None:
-        """
-        Visits augmented assigns.
-
-        Raises:
-            ZeroDivisionViolation
-
-        """
-        self._check_zero_division(node.op, node.value)
-        self._check_useless_math_operator(node.op, node.value)
-        self.generic_visit(node)
-
-    def _check_operator_count(self, node: ast.Num) -> None:
-        for node_type, limit in self._limits.items():
-            if count_unary_operator(node, node_type) > limit:
-                self.add_violation(
-                    consistency.UselessOperatorsViolation(
-                        node, text=str(node.n),
-                    ),
-                )
-
-    def _check_zero_division(self, op: ast.operator, number: ast.AST) -> None:
-        number = unwrap_unary_node(number)
-
-        is_zero_division = (
-            isinstance(op, ast.Div) and
-            isinstance(number, ast.Num) and
-            number.n == 0
-        )
-        if is_zero_division:
-            self.add_violation(consistency.ZeroDivisionViolation(number))
-
-    def _check_useless_math_operator(
-        self,
-        op: ast.operator,
-        left: ast.AST,
-        right: Optional[ast.AST] = None,
-    ) -> None:
-        non_negative_numbers = self._get_non_negative_nodes(left, right)
-
-        for number in non_negative_numbers:
-            forbidden = self._meaningless_operations.get(number.n, None)
-            if forbidden and isinstance(op, forbidden):
-                self.add_violation(
-                    consistency.MeaninglessNumberOperationViolation(number),
-                )
-
-    def _get_non_negative_nodes(
-        self,
-        left: ast.AST,
-        right: Optional[ast.AST] = None,
-    ):
-        non_negative_numbers = []
-        for node in filter(None, (left, right)):
-            real_node = unwrap_unary_node(node)
-            if not isinstance(real_node, ast.Num):
-                continue
-
-            if real_node.n not in self._meaningless_operations:
-                continue
-
-            if real_node.n == 1 and walk.is_contained(node, ast.USub):
-                continue
-            non_negative_numbers.append(real_node)
-        return non_negative_numbers
-
-
-@final
 @decorators.alias('visit_any_for', (
     'visit_For',
     'visit_AsyncFor',
@@ -267,6 +143,9 @@ class WrongAssignmentVisitor(base.BaseNodeVisitor):
     def visit_Assign(self, node: ast.Assign) -> None:
         """
         Checks assignments to be correct.
+
+        We do not check ``AnnAssign`` here,
+        because it does not have problems that we check.
 
         Raises:
             MultipleAssignmentsViolation